Extract DynamoError class
