+++ app/services/DynamoChannelTests.scala
@@ -4,8 +4,7 @@ import com.typesafe.scalalogging.StrictLogging
 import io.circe.{Decoder, Encoder}
 import io.circe.syntax._
 import io.circe.generic.auto._
-import models.{Channel, ChannelTest, LockStatus, Status}
-import DynamoChannelTests._
+import models.{Channel, ChannelTest, DynamoDuplicateNameError, DynamoError, DynamoGetError, DynamoNoLockError, DynamoPutError, LockStatus, Status}
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient
 import software.amazon.awssdk.services.dynamodb.model._
 import utils.Circe.{dynamoMapToJson, jsonToDynamo}
@@ -16,22 +15,6 @@ import zio.stream.ZStream
 
 import java.time.OffsetDateTime
 import scala.jdk.CollectionConverters._
-
-object DynamoChannelTests {
-  sealed trait DynamoError extends Throwable
-  case class DynamoPutError(error: Throwable) extends DynamoError {
-    override def getMessage = s"Error writing to Dynamo: ${error.getMessage}"
-  }
-  case class DynamoNoLockError(error: Throwable) extends DynamoError {
-    override def getMessage = s"Error writing to Dynamo: ${error.getMessage}"
-  }
-  case class DynamoGetError(error: Throwable) extends DynamoError {
-    override def getMessage = s"Error reading from Dynamo: ${error.getMessage}"
-  }
-  case class DynamoDuplicateNameError(error: Throwable) extends DynamoError {
-    override def getMessage = s"Error writing to Dynamo: ${error.getMessage}"
-  }
-}
 
 class DynamoChannelTests(stage: String, client: DynamoDbClient) extends StrictLogging {
